Rename lemur_azure_dest package to lemur_azure
