+++ publish/files/disable_plugins.patch
@@ -32,7 +32,7 @@
 -            'adcs_source = lemur.plugins.lemur_adcs.plugin:ADCSSourcePlugin',
 -            'entrust_issuer = lemur.plugins.lemur_entrust.plugin:EntrustIssuerPlugin',
 -            'entrust_source = lemur.plugins.lemur_entrust.plugin:EntrustSourcePlugin',
-             'azure_destination = lemur.plugins.lemur_azure_dest.plugin:AzureDestinationPlugin'
+             'azure_destination = lemur.plugins.lemur_azure.plugin:AzureDestinationPlugin'
          ],
      },
      classifiers=[
