+++ setup.py
@@ -158,7 +158,7 @@ setup(
             'adcs_source = lemur.plugins.lemur_adcs.plugin:ADCSSourcePlugin',
             'entrust_issuer = lemur.plugins.lemur_entrust.plugin:EntrustIssuerPlugin',
             'entrust_source = lemur.plugins.lemur_entrust.plugin:EntrustSourcePlugin',
-            'azure_destination = lemur.plugins.lemur_azure_dest.plugin:AzureDestinationPlugin',
+            'azure_destination = lemur.plugins.lemur_azure.plugin:AzureDestinationPlugin',
             'gcp_destination = lemur.plugins.lemur_gcp.plugin:GCPDestinationPlugin',
             'gcp_source = lemur.plugins.lemur_gcp.plugin:GCPSourcePlugin'
         ],
